LayoutTests/platform/ios-simulator/transitions/transform-op-list-no-match-expected.txt

CONSOLE MESSAGE: line 240: Failed to pause '-webkit-transform' transition on element 'box'
CONSOLE MESSAGE: line 240: Failed to pause '-webkit-transform' transition on element 'box'
Box should spin only 180 degrees even though the operation specifies a 0 to 540 degree animation. This is because the operation lists don't match, so it falls back to matrix animation, which normalizes the angle to between 0 and 360 degrees.

FAIL - "-webkit-transform.0" property for "box" element at 0.5s expected: 0.70711 but saw: matrix(-1, 0.00000000000000036739403974420594, -0.00000000000000036739403974420594, -1, 0, 0)
FAIL - "-webkit-transform.0" property for "box" element at 1s expected: 0 but saw: matrix(-1, 0.00000000000000036739403974420594, -0.00000000000000036739403974420594, -1, 0, 0)
